+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.0.1
+26 may 2017
+
+* Function duplicate is removed
+* encrypt can create folder if it's necessary
+* decrypt now can create folders and rewrite files if it's necessary
+* fixed multi callback in decrypt
+
 ## v1.0.0
 25 may 2017